add Products() adminMenu

+++ Products/Clothing/Clothing.cs
@@ -12,14 +12,14 @@ namespace GarbageCollectorStore
         public string Material {  get; set; }
         public List<string> Sizes { get; set; }
 
-        public Clothing(string material, string name, double price, int quantity, string description, List<string> sizes)
-            : base(name, (decimal)price, quantity, description)
+        public Clothing(string material, string name, decimal price, int quantity, string description, List<string> sizes)
+            : base(name, price, quantity, description)
         {
             Material = material;
             Sizes = new List<string>(sizes);
         }
-        public Clothing(string material, string name, double price, int quantity, string description, params string[]  sizes)
-    : base(name, (decimal)price, quantity, description)
+        public Clothing(string material, string name, decimal price, int quantity, string description, params string[]  sizes)
+    : base(name, price, quantity, description)
         {
             Material = material;
             Sizes = new List<string>(sizes);
